Create getcheckhash.sh
Manage file downloads across many thousands of devices. Wait a random amount of seconds to get the file as not to overwhelm the http server, than check the sha1sum against the passed hash. If the hashes do not match, download the file again. Do this up to five times, sleeping for a random interval between attempts.